+++ core/src/main/java/nl/nn/adapterframework/http/rest/ApiEhcache.java
@@ -116,6 +116,12 @@ public class ApiEhcache implements IApiCache {
 
 
 	public Object get(String key) {
+		// Workaround to avoid NPE after a full reload (/adapterHandlerAsAdmin.do?action=fullreload)
+		// get() and isKeyInCache() are not synchronized methods and do not contain any state checking.
+		// The cache can only check if a key exists if it's state is ALIVE.
+		if(!cache.getStatus().equals(Status.STATUS_ALIVE))
+			return null;
+
 		Element element = cache.get(key);
 		if (element==null) {
 			return null;
@@ -138,14 +144,8 @@ public class ApiEhcache implements IApiCache {
 		return cache.remove(key);
 	}
 
-	public boolean containsKey(String string) {
-		//workaround to avoid NPE after a full reload (/adapterHandlerAsAdmin.do?action=fullreload)
-		//TODO: fix this in a proper way
-		if (cache.getStatus().equals(Status.STATUS_SHUTDOWN)) {
-			log.info("cache ["+KEY_CACHE_NAME+"] has status shutdown, so returning false");
-			return false;
-		}
-		return cache.isKeyInCache(string);
+	public boolean containsKey(String key) {
+		return (this.get(key) != null);
 	}
 
 	public void flush() {